requirements.txt
discord.py==2.3.2
openai_whisper==20231117
python-dotenv==1.0.1
whisper==1.1.10
